+++ project_name/urls.py
@@ -5,36 +5,47 @@ The `urlpatterns` list routes URLs to views.
 
 For more information on this file, see:
 https://docs.djangoproject.com/en/{{ docs_version }}/topics/http/urls/
-
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function:  from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 
+from django.conf import settings
+from django.conf.urls.i18n import i18n_patterns
+from django.conf.urls.static import static
 from django.contrib import admin
-from django.urls import path
+from django.urls import include, path
+from django.views import defaults as default_views
 
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
+    path('i18n/', include('django.conf.urls.i18n')),
 ]
 
-# Internationalization
-# https://docs.djangoproject.com/en/{{ docs_version }}/topics/i18n/
-# from django.conf.urls.i18n import i18n_patterns
-# from django.utils.translation import gettext_lazy as _
-# urlpatterns += i18n_patterns()
+urlpatterns += i18n_patterns(
+    path('admin/', admin.site.urls),
+)
 
-# Serve files uploaded by users during development
-# https://docs.djangoproject.com/en/{{ docs_version }}/howto/static-files/
-# from django.conf import settings
-# from django.conf.urls.static import static
-# if settings.DEBUG:
-#     urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+if settings.DEBUG:
+    # Browse custom error pages during development
+    urlpatterns += i18n_patterns(
+        path(
+            '400/',
+            default_views.bad_request,
+            kwargs={'exception': Exception('Bad Request')},
+        ),
+        path(
+            '403/',
+            default_views.permission_denied,
+            kwargs={'exception': Exception('Forbidden')},
+        ),
+        path(
+            '404/',
+            default_views.page_not_found,
+            kwargs={'exception': Exception('Not Found')},
+        ),
+        path(
+            '500/',
+            default_views.server_error,
+        ),
+    )
+    # Serve files uploaded by users during development
+    urlpatterns += static(
+        settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)